+++ app/src/functions/qconcrusos/questions/runlocal.ts
@@ -16,7 +16,7 @@ const event = {
             "MessageId": "67d2b442-ef5c-4e22-b1a5-3e18ef25e664",
             "TopicArn": "arn:aws:sns:sa-east-1:11111:test",
             "Subject": null,
-            "Message": "{\n  \"url\": \"https://www.qconcursos.com/questoes-de-concursos/questoes?institute_ids%5B%5D=20&knowledge_area_ids%5B%5D=10&page=38\"\n}",
+            "Message": "https://www.qconcursos.com/questoes-de-concursos/questoes?institute_ids%5B%5D=20&knowledge_area_ids%5B%5D=10&page=38",
             "Timestamp": "2022-01-08T03:44:12.824Z",
             "SignatureVersion": "1",
             "Signature": "",
